+++ griptape/schemas/base_schema.py
@@ -2,12 +2,14 @@ from __future__ import annotations
 
 from abc import ABC
 from collections.abc import Sequence
+from enum import Enum
 from typing import Any, Literal, TypeVar, Union, _SpecialForm, get_args, get_origin
 
 import attrs
 from marshmallow import INCLUDE, Schema, fields
 
 from griptape.schemas.bytes_field import Bytes
+from griptape.schemas.union_field import Union as UnionField
 
 
 class BaseSchema(Schema):
@@ -59,21 +61,63 @@ class BaseSchema(Schema):
         # Resolve TypeVars to their bound type
         if isinstance(field_class, TypeVar):
             field_class = field_class.__bound__
-
-        if attrs.has(field_class):
-            if ABC in field_class.__bases__:
-                return fields.Nested(PolymorphicSchema(inner_class=field_class), allow_none=optional)
-            else:
-                return fields.Nested(cls.from_attrs_cls(field_class), allow_none=optional)
-        elif cls.is_list_sequence(field_class):
+        if field_class is None:
+            return fields.Constant(None, allow_none=True)
+        if cls._is_union(field_type):
+            return cls._handle_union(field_type, optional=optional)
+        elif attrs.has(field_class):
+            schema = PolymorphicSchema if ABC in field_class.__bases__ else cls.from_attrs_cls
+            return fields.Nested(schema(field_class), allow_none=optional)
+        elif cls._is_enum(field_type):
+            return fields.String(allow_none=optional)
+        elif cls._is_list_sequence(field_class):
             if args:
-                return fields.List(cls_or_instance=cls._get_field_for_type(args[0]), allow_none=optional)
+                return cls._handle_list(args[0], optional=optional)
             else:
                 raise ValueError(f"Missing type for list field: {field_type}")
-        else:
-            field_class = cls.DATACLASS_TYPE_MAPPING[field_class]
+        field_class = cls.DATACLASS_TYPE_MAPPING.get(field_class)
+        if field_class is None:
+            raise ValueError(f"Unsupported field type: {field_type}")
+        return field_class(allow_none=optional)
 
-            return field_class(allow_none=optional)
+    @classmethod
+    def _handle_list(cls, list_type: type, *, optional: bool) -> fields.Field:
+        """Handle List Fields, including Union Types.
+
+        Args:
+            list_type: The List type to handle.
+            optional: Whether the List can be none.
+
+        Returns:
+            A marshmallow List field.
+        """
+        if cls._is_union(list_type):
+            union_field = cls._handle_union(list_type, optional=optional)
+            return fields.List(cls_or_instance=union_field, allow_none=optional)
+        list_field = cls._get_field_for_type(list_type)
+        if isinstance(list_field, fields.Constant) and list_field.constant is None:
+            raise ValueError(f"List elements cannot be None: {list_type}")
+        return fields.List(cls_or_instance=list_field, allow_none=optional)
+
+    @classmethod
+    def _handle_union(cls, union_type: type, *, optional: bool) -> fields.Field:
+        """Handle Union Fields, including Unions with List Types.
+
+        Args:
+            union_type: The Union Type to handle.
+            optional: Whether the Union can be None.
+
+        Returns:
+            A marshmallow Union field.
+        """
+        candidate_fields = [cls._get_field_for_type(arg) for arg in get_args(union_type) if arg is not type(None)]
+        optional_args = [arg is None for arg in get_args(union_type)]
+        if optional_args:
+            optional = True
+        if not candidate_fields:
+            raise ValueError(f"Unsupported UnionType field: {union_type}")
+
+        return UnionField(fields=candidate_fields, allow_none=optional)
 
     @classmethod
     def _get_field_type_info(cls, field_type: type) -> tuple[type, tuple[type, ...], bool]:
@@ -178,7 +222,7 @@ class BaseSchema(Schema):
         )
 
     @classmethod
-    def is_list_sequence(cls, field_type: type | _SpecialForm) -> bool:
+    def _is_list_sequence(cls, field_type: type | _SpecialForm) -> bool:
         if isinstance(field_type, type):
             if issubclass(field_type, str) or issubclass(field_type, bytes) or issubclass(field_type, tuple):
                 return False
@@ -186,3 +230,11 @@ class BaseSchema(Schema):
                 return issubclass(field_type, Sequence)
         else:
             return False
+
+    @classmethod
+    def _is_union(cls, field_type: type) -> bool:
+        return field_type is Union or get_origin(field_type) is Union
+
+    @classmethod
+    def _is_enum(cls, field_type: type) -> bool:
+        return isinstance(field_type, type) and issubclass(field_type, Enum)